use filesystem pattern for apis
